backend/requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
bitarray==3.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.2
cytoolz==1.0.1
defusedxml==0.7.1
Django==5.1.6
django-allauth==65.4.1
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
hexbytes==1.3.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
multidict==6.1.0
oauthlib==3.2.2
parsimonious==0.10.0
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PyNaCl==1.5.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pyunormalize==16.0.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.1.0
rpds-py==0.23.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
ton-client-py==1.44.3.0
tonsdk==1.0.15
toolz==1.0.0
types-requests==2.32.0.20250301
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
web3==7.8.0
websockets==13.1
yarl==1.18.3
